basic mad-libs game in Python
In this example, the story variable contains a predefined story with placeholders in the form of curly braces {}. The play_madlibs function prompts the player to enter various words such as adjectives, nouns and verbs. These words are then inserted into the story using the .format() method and the placeholders in the story are replaced with the input words. The final story with the input words is printed to the screen.

This is just a basic example, you can add more placeholders to the story, or you can have multiple stories and have the player choose which story they want to play with.

You can also make the game more interactive by adding more features like score keeping, or different levels of difficulty.